добавил ридми
добавил readOnly
