+++ readme.md
@@ -1,0 +1,46 @@
+### Описание задачи
+
+Необходимо разработать микросервис на Spring Boot 3, который будет
+предоставлять REST API для управления пользователями и их подписками на
+сервисы.
+Проект использует Java 17
+
+### Функциональные требования
+
+1 API для управления пользователями
+    - Создание пользователя.
+    - Получение информации о пользователе.
+    - Обновление данных пользователя.
+    - Удаление пользователя.
+
+
+2 API для подписок
+    - Добавление подписки пользователю.
+    - Получение списка подписок пользователя.
+    - Удаление подписки.
+    - Подписки представляют собой подписки на цифровые сервисы, такие как
+        YouTube Premium, VK Музыка, Яндекс.Плюс, Netflix и другие стриминговые
+    платформы.
+3 Интеграция с базой данных
+    - Использовать PostgreSQL.
+    - Таблицы: users, subscriptions.
+
+4 Логирование
+    - Логирование через SLF4J.
+
+5 Docker
+  - Создать Dockerfile для развертывания сервиса.
+  - Разработать docker-compose.yml, который позволит локально запускать проект
+    вместе с базой данных.
+
+### Требования к API
+
+Примерные эндпоинты:
+POST /users - создать пользователя
+GET /users/{id} - получить информацию о пользователе
+PUT /users/{id} - обновить пользователя
+DELETE /users/{id} - удалить пользователя
+POST /users/{id}/subscriptions - добавить подписку
+GET /users/{id}/subscriptions - получить подписки пользователя
+DELETE /users/{id}/subscriptions/{sub_id} - удалить подписку
+GET /subscriptions/top - получить ТОП-3 популярных подписок
