+++ src/main/java/org/example/controller/SubscriptionController.java
@@ -5,10 +5,10 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
-import org.example.responses.CommonErrorApiResponses;
 import org.example.dto.request.RequestSubscriptionDTO;
 import org.example.dto.response.ResponseSubscriptionDTO;
-import org.example.service.impl.SubscriptionServiceImpl;
+import org.example.responses.CommonErrorApiResponses;
+import org.example.service.SubscriptionService;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,7 +20,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SubscriptionController {
 
-    private final SubscriptionServiceImpl subscriptionServiceImpl;
+    private final SubscriptionService subscriptionService;
 
     @Operation(summary = "Получение топ 3 подписок по кол-ву пользователей")
     @ApiResponse(responseCode = "200", description = "Успешное выполнение операции",
@@ -29,7 +29,7 @@ public class SubscriptionController {
     @CommonErrorApiResponses
     @GetMapping("/top")
     public ResponseEntity<List<ResponseSubscriptionDTO>> getTopOfSubscriptions() {
-        return ResponseEntity.ok(subscriptionServiceImpl.getTopOfSubscriptions());
+        return ResponseEntity.ok(subscriptionService.getTopOfSubscriptions());
     }
 
 
@@ -40,6 +40,6 @@ public class SubscriptionController {
     @CommonErrorApiResponses
     @PostMapping
     public ResponseEntity<ResponseSubscriptionDTO> createSubscription(@RequestBody RequestSubscriptionDTO requestSubscriptionDTO) {
-        return ResponseEntity.ok(subscriptionServiceImpl.createSubscription(requestSubscriptionDTO));
+        return ResponseEntity.ok(subscriptionService.createSubscription(requestSubscriptionDTO));
     }
 }
