+++ src/main/java/org/example/service/impl/UserServiceImpl.java
@@ -1,17 +1,15 @@
 package org.example.service.impl;
 
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.example.responses.Errors;
 import org.example.dto.entity.User;
 import org.example.dto.request.RequestUserDTO;
 import org.example.dto.response.ResponseUserDTO;
 import org.example.exception.extend.user.UserNotFoundException;
-import org.example.exception.extend.user.UsernameTakenException;
 import org.example.mapper.UserMapper;
 import org.example.repository.UserRepository;
 import org.example.service.UserService;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,9 +17,11 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepo;
     private final UserMapper userMapper;
+    private final UserUtilService userUtilService;
 
 
     @Override
+    @Transactional(readOnly = true)
     public ResponseUserDTO getUserById(Long id) {
         User user = userRepo.findById(id).orElseThrow(() -> new UserNotFoundException("User not found"));
         return userMapper.toResponseUserDTO(user);
@@ -31,7 +31,7 @@ public class UserServiceImpl implements UserService {
     @Transactional
     public ResponseUserDTO createUser(RequestUserDTO requestUserDTO) {
         User user = new User();
-        setUserParams(user, requestUserDTO);
+        userUtilService.setUserParams(user, requestUserDTO);
 
         userRepo.save(user);
         return userMapper.toResponseUserDTO(user);
@@ -41,7 +41,7 @@ public class UserServiceImpl implements UserService {
     @Transactional
     public ResponseUserDTO updateUser(Long id, RequestUserDTO requestUserDTO) {
         User user = userRepo.findById(id).orElseThrow(() -> new UserNotFoundException("User not found"));
-        setUserParams(user, requestUserDTO);
+        userUtilService.setUserParams(user, requestUserDTO);
         userRepo.save(user);
         return userMapper.toResponseUserDTO(user);
     }
@@ -51,26 +51,4 @@ public class UserServiceImpl implements UserService {
         userRepo.deleteById(id);
     }
 
-    private void setUserParams(User user, RequestUserDTO requestUserDTO) {
-        if (validateUsername(requestUserDTO.getUsername(), user.getId())) {
-            throw new UsernameTakenException("Username is taken", Errors.USERNAME_TAKEN);
-        }
-        if (validateEmail(requestUserDTO.getEmail(), user.getId())) {
-            throw new UsernameTakenException("Email is invalid or taken", Errors.EMAIL_TAKEN);
-        }
-        user.setUsername(requestUserDTO.getUsername());
-        user.setEmail(requestUserDTO.getEmail());
-        user.setFirstname(requestUserDTO.getFirstname());
-        user.setLastname(requestUserDTO.getLastname());
-        user.setMiddlename(requestUserDTO.getMiddlename());
-
-    }
-
-    private boolean validateUsername(String username, Long id) {
-        return userRepo.existsByUsernameAndIdNot(username, id);
-    }
-
-    private boolean validateEmail(String email, Long id) {
-        return userRepo.existsByEmailAndIdNot(email, id);
-    }
 }
